Fixed connectivity issues.
The data engine now monitors the state of the network and will
attempt to (re-)connect at regular intervals if the network is
down. Similarly, it will close and try to reopen the connection
if it detects that data has not been arriving for a while (this
occurs only if regular updates are being requested).